pass type to build_simulation_grid_axis

+++ workflow/src/LegendSimflow.jl/test/test_lib.jl
@@ -66,7 +66,7 @@ end
     meta_path = normpath(joinpath(@__DIR__, "..", "..", "..", "..", "tests", "dummyprod", "inputs"))
     det = "V99000A"
     opv_val = 3000.0
-    T = Float64
+    T = Float32
     refinement_limits = [0.2, 0.1, 0.05, 0.02]
 
     meta, xtal, opv = load_detector_metadata(meta_path, det, opv_val)
